feat: convert R.string to Res.string and move VoicePreset + TorSettings to commons
Move VoicePreset enum and TorSettings/TorType/TorPresetType to commons module,
converting Int resource IDs to StringResource from Compose Multiplatform Resources.

- VoicePreset: labelRes changed from Int to StringResource
- TorType: resourceId changed from Int to StringResource
- TorPresetType: resourceId/explainerId changed from Int to StringResource
- Added voice_preset_* and tor_* strings to commons strings.xml
- Changed compose-resources from implementation to api in commons
- Updated VoicePresetSelector and TorSettingsDialog callers
- Left backward-compat typealiases at original locations

Part of KMP iOS migration (#2238).
